Refactor ptb_jobstores for separation of modules & dependencies
