+++ ptbcontrib/ptb_jobstores/sqlalchemy/__init__.py
@@ -1,0 +1,20 @@
+# A library containing community-based extension for the python-telegram-bot library
+# Copyright (C) 2020-2023
+# The ptbcontrib developers
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser Public License for more details.
+#
+# You should have received a copy of the GNU Lesser Public License
+# along with this program.  If not, see [http://www.gnu.org/licenses/].
+"""This module contains PTB SQLAlchemy JobStore."""
+from .sqlalchemy import PTBSQLAlchemyJobStore
+
+__all__ = ["PTBSQLAlchemyJobStore"]
